reorganise calendar services
