+++ packages/twenty-server/src/modules/calendar/calendar-event-participant-manager/jobs/calendar-event-participant-unmatch-participant.job.ts
@@ -3,7 +3,7 @@ import { Scope } from '@nestjs/common';
 import { Processor } from 'src/engine/integrations/message-queue/decorators/processor.decorator';
 import { MessageQueue } from 'src/engine/integrations/message-queue/message-queue.constants';
 import { Process } from 'src/engine/integrations/message-queue/decorators/process.decorator';
-import { CalendarEventParticipantService } from 'src/modules/calendar/common/services/calendar-event-participant.service';
+import { CalendarEventParticipantService } from 'src/modules/calendar/calendar-event-participant-manager/services/calendar-event-participant.service';
 
 export type CalendarEventParticipantUnmatchParticipantJobData = {
   workspaceId: string;
